+++ src/main/java/com/aoyama/modernfont/AoyamasModernFont.java
@@ -36,7 +36,7 @@ public class AoyamasModernFont {
      * MODバージョン。
      */
     public static final String VERSION =
-            "1.0.0";
+            "1.0.1";
 
     /**
      * フォントの検索と読み込みを管理するFontManager。